+++ app/uploaders/evss_claim_document_uploader.rb
@@ -5,31 +5,7 @@ class EVSSClaimDocumentUploader < CarrierWave::Uploader::Base
   include CarrierWave::MiniMagick
   include SetAWSConfig
   include ValidateEVSSFileSize
-
-  class << self
-    def tiff?(mimemagic_object: nil, carrier_wave_sanitized_file: nil)
-      return mimemagic_object.type == 'image/tiff' if mimemagic_object
-
-      carrier_wave_sanitized_file&.content_type == 'image/tiff'
-    end
-
-    def incorrect_extension?(extension:, mimemagic_object:)
-      extension = extension.to_s.downcase
-      true_extensions = extensions_from_mimemagic_object(mimemagic_object).map(&:downcase)
-      true_extensions.present? && !extension.in?(true_extensions)
-    end
-
-    def extensions_from_mimemagic_object(mimemagic_object)
-      mimemagic_object&.extensions || []
-    end
-
-    def inspect_binary(carrier_wave_sanitized_file)
-      file_obj = carrier_wave_sanitized_file&.to_file
-      file_obj && MimeMagic.by_magic(file_obj)
-    ensure
-      file_obj.close if file_obj.respond_to? :close
-    end
-  end
+  include ConvertFileType
 
   def size_range
     1.byte...150.megabytes
@@ -37,7 +13,6 @@ class EVSSClaimDocumentUploader < CarrierWave::Uploader::Base
 
   version :converted, if: :tiff_or_incorrect_extension? do
     process(convert: :jpg, if: :tiff?)
-
     def full_filename(original_name_for_file)
       name = "converted_#{original_name_for_file}"
       extension = CarrierWave::SanitizedFile.new(nil).send(:split_extension, original_name_for_file)[1]
@@ -58,18 +33,6 @@ class EVSSClaimDocumentUploader < CarrierWave::Uploader::Base
     @user_uuid = user_uuid
     @ids = ids
     set_storage_options!
-  end
-
-  def converted_exists?
-    converted.present? && converted.file.exists?
-  end
-
-  def final_filename
-    if converted_exists?
-      converted.file.filename
-    else
-      filename
-    end
   end
 
   def read_for_upload
@@ -97,24 +60,6 @@ class EVSSClaimDocumentUploader < CarrierWave::Uploader::Base
   end
 
   private
-
-  def tiff?(carrier_wave_sanitized_file)
-    self.class.tiff?(
-      carrier_wave_sanitized_file: carrier_wave_sanitized_file,
-      mimemagic_object: self.class.inspect_binary(carrier_wave_sanitized_file)
-    )
-  end
-
-  def tiff_or_incorrect_extension?(carrier_wave_sanitized_file)
-    mimemagic_object = self.class.inspect_binary carrier_wave_sanitized_file
-    self.class.tiff?(
-      carrier_wave_sanitized_file: carrier_wave_sanitized_file,
-      mimemagic_object: mimemagic_object
-    ) || self.class.incorrect_extension?(
-      extension: carrier_wave_sanitized_file.extension,
-      mimemagic_object: mimemagic_object
-    )
-  end
 
   def set_storage_options!
     if Settings.evss.s3.uploads_enabled
